+++ Applet/Package.swift
@@ -10,6 +10,7 @@ let package = Package(
         .executable(name: "NoodletPreview", targets: ["NoodletPreview"]),
     ],
     dependencies: [
+        .package(path: "../Shared/SettingsUI"),
         .package(path: "Protocol"),
         .package(path: "../Shared/Wallpaper"),
         .package(url: "https://github.com/sparkle-project/Sparkle", exact: "2.9.4"),
@@ -21,6 +22,7 @@ let package = Package(
             name: "NoodleApplet",
             dependencies: [
                 "AppletCore", .product(name: "AppletBridge", package: "Protocol"),
+                .product(name: "NoodleSettingsUI", package: "SettingsUI"),
                 .product(name: "NoodleWallpaper", package: "Wallpaper"),
                 .product(name: "Sparkle", package: "Sparkle"),
             ],